+++ unit-tests/forms/disable-save/power.test.js
@@ -217,4 +217,161 @@ describe("power", () => {
       assert.isTrue(actualData, "it should be disabled");
     });
   });
+  describe("power vs ssh keys", () => {
+    it("should return true when the ssh key name already exists", () => {
+      assert.isTrue(
+        disableSave(
+          "ssh_keys",
+          {
+            name: "honk",
+            resource_group: "hi",
+            public_key:
+              "ssh-rsa AAAAB3NzaC1yc2thisisafakesshkeyDSKLFHSJSADFHGASJDSHDBASJKDASDASWDAS+/DSFSDJKFGXFVJDZHXCDZVZZCDKJFGSDJFZDHCVBSDUCZCXZKCHT= test@fakeemail.com",
+          },
+          {
+            data: {
+              data: "test",
+            },
+            arrayParentName: "workspace",
+            craig: {
+              store: {
+                resourceGroups: ["hi"],
+                json: {
+                  ssh_keys: [],
+                  power: [
+                    {
+                      name: "workspace",
+                      ssh_keys: [
+                        {
+                          name: "honk",
+                        },
+                      ],
+                    },
+                  ],
+                },
+              },
+            },
+          }
+        ),
+        "it should be disabled"
+      );
+    });
+    it("should return true when the ssh key has no rg", () => {
+      assert.isTrue(
+        disableSave(
+          "ssh_keys",
+          {
+            name: "honk",
+            resource_group: null,
+            public_key:
+              "ssh-rsa AAAAB3NzaC1yc2thisisafakesshkeyDSKLFHSJSADFHGASJDSHDBASJKDASDASWDAS+/DSFSDJKFGXFVJDZHXCDZVZZCDKJFGSDJFZDHCVBSDUCZCXZKCHT= test@fakeemail.com",
+          },
+          {
+            data: {
+              data: "test",
+            },
+            arrayParentName: "workspace",
+            craig: {
+              store: {
+                resourceGroups: ["hi"],
+                json: {
+                  ssh_keys: [],
+                  power: [
+                    {
+                      name: "workspace",
+                      ssh_keys: [
+                        {
+                          name: "ddd",
+                        },
+                      ],
+                    },
+                  ],
+                },
+              },
+            },
+          }
+        ),
+        "it should be disabled"
+      );
+    });
+    it("should return true when the public key value already exists", () => {
+      assert.isTrue(
+        disableSave(
+          "ssh_keys",
+          {
+            name: "test",
+            resource_group: "hi",
+            public_key:
+              "ssh-rsa AAAAB3NzaC1yc2thisisafakesshkeyDSKLFHSJSADFHGASJDSHDBASJKDASDASWDAS+/DSFSDJKFGXFVJDZHXCDZVZZCDKJFGSDJFZDHCVBSDUCZCXZKCHT= test@fakeemail.com",
+          },
+          {
+            data: {
+              data: "test",
+            },
+            arrayParentName: "workspace",
+            craig: {
+              store: {
+                resourceGroups: ["hi"],
+                json: {
+                  ssh_keys: [
+                    {
+                      name: "honk",
+                      public_key:
+                        "ssh-rsa AAAAB3NzaC1yc2thisisafakesshkeyDSKLFHSJSADFHGASJDSHDBASJKDASDASWDAS+/DSFSDJKFGXFVJDZHXCDZVZZCDKJFGSDJFZDHCVBSDUCZCXZKCHT= test@fakeemail.com",
+                    },
+                  ],
+                  power: [
+                    {
+                      name: "workspace",
+                      ssh_keys: [
+                        {
+                          name: "ddd",
+                          public_key:
+                            "ssh-rsa AAAAB3NzaC1yc2thisisafakesshkeyDSKLFHSJSADFHGASJDSHDBASJKDASDASWDAS+/DSFSDJKFGXFVJDZHXCDZVZZCDKJFGSDJFZDHCVBSDUCZCXZKCHT= test@fakeemail.com",
+                        },
+                      ],
+                    },
+                  ],
+                },
+              },
+            },
+          }
+        ),
+        "it should be disabled"
+      );
+    });
+    it("should not check invalidSshKey when using data", () => {
+      assert.isFalse(
+        disableSave(
+          "ssh_keys",
+          {
+            name: "test",
+            resource_group: "hi",
+            public_key: "honk",
+            use_data: true,
+          },
+          {
+            data: {
+              data: "test",
+            },
+            craig: {
+              store: {
+                resourceGroups: ["hi"],
+                json: {
+                  ssh_keys: [
+                    {
+                      name: "honk",
+                      public_key:
+                        "ssh-rsa AAAAB3NzaC1yc2thisisafakesshkeyDSKLFHSJSADFHGASJDSHDBASJKDASDASWDAS+/DSFSDJKFGXFVJDZHXCDZVZZCDKJFGSDJFZDHCVBSDUCZCXZKCHT= test@fakeemail.com",
+                    },
+                  ],
+                },
+              },
+            },
+          }
+        ),
+        "it should be enabled"
+      );
+    });
+  });
 });